Crate uniffi_macros

Macros for uniffi.

Currently this is just for easily generating integration tests, but maybe we’ll put some other code-annotation helper macros in here at some point.

Macros

build_foreign_language_testcases
A macro to build testcases for a component’s generated bindings.

generate_and_include_scaffolding
A helper macro to generate and include component scaffolding.

include_scaffolding
A helper macro to include generated component scaffolding.

Structs

FilePaths
Newtype to simplifying parsing a list of file paths from macro input.

Functions

should_skip_path