Struct uniffi_macros::export::scaffolding::ScaffoldingBits

struct ScaffoldingBits {
    param_names: Vec<TokenStream>,
    param_types: Vec<TokenStream>,
    lift_closure: TokenStream,
    rust_fn_call: TokenStream,
    convert_result: TokenStream,
}

Fields
param_names: Vec<TokenStream>

Parameter names for the scaffolding function

param_types: Vec<TokenStream>

Parameter types for the scaffolding function

lift_closure: TokenStream

Lift closure. See FnSignature::lift_closure for an explanation of this.

rust_fn_call: TokenStream

Expression to call the Rust function after a successful lift.

convert_result: TokenStream

Convert the result of rust_fn_call, stored in a variable named uniffi_result into its final value. This is used to do things like error conversion / Arc wrapping

Implementations

impl ScaffoldingBits

fn new_for_function(sig: &FnSignature, udl_mode: bool) -> Self

fn new_for_method( sig: &FnSignature, self_ident: &Ident, is_trait: bool, udl_mode: bool ) -> Self

fn new_for_constructor( sig: &FnSignature, self_ident: &Ident, udl_mode: bool ) -> Self