Macro uniffi_macros::generate_and_include_scaffolding

generate_and_include_scaffolding!() { /* proc-macro */ }

A helper macro to generate and include component scaffolding.

This is a convenience macro designed for writing trybuild-style tests and probably shouldn’t be used for production code. Given the path to a .udl file, if will run uniffi-bindgen to produce the corresponding Rust scaffolding and then include it directly into the calling file. Like so:

uniffi_macros::generate_and_include_scaffolding!("path/to/my/interface.udl");