Crate uniffi_macros

Macros for uniffi.

Currently this is just for easily generating integration tests, but maybe we’ll put some other code-annotation helper macros in here at some point.

Modules

enum_ 🔒

error 🔒

export 🔒

object 🔒

record 🔒

test 🔒

util 🔒

Macros

build_foreign_language_testcases

A macro to build testcases for a component’s generated bindings.

generate_and_include_scaffolding

A helper macro to generate and include component scaffolding.

include_scaffolding

A helper macro to include generated component scaffolding.

Attribute Macros

export

Derive Macros

Enum

Error

Object

Record