Crate uniffi_macros

Macros for uniffi.

Currently this is just for easily generating integration tests, but maybe we’ll put some other code-annotation helper macros in here at some point.

Modules

custom 🔒

enum_ 🔒

error 🔒

export 🔒

fnsig 🔒

object 🔒

record 🔒

setup_scaffolding 🔒

test 🔒

util 🔒

Macros

build_foreign_language_testcases

A macro to build testcases for a component’s generated bindings.

custom_newtype

Generate the FfiConverter and the UniffiCustomTypeConverter implementations for a Custom Type - ie, for a <T> which implements UniffiCustomTypeConverter via the newtype idiom.

custom_type

Generate the FfiConverter implementation for a Custom Type - ie, for a <T> which implements UniffiCustomTypeConverter.

generate_and_include_scaffolding

A helper macro to generate and include component scaffolding.

include_scaffolding

A helper macro to include generated component scaffolding.

setup_scaffolding

Top-level initialization macro

use_udl_enum

use_udl_error

use_udl_object

use_udl_record

Structs

CustomTypeInfo 🔒

Functions

do_export 🔒

use_udl_simple_type 🔒

Attribute Macros

constructor

A dummy macro that does nothing.

export

Derive Macros

Enum

Error

Object

Record