Struct uniffi_macros::export::scaffolding::ScaffoldingBits

struct ScaffoldingBits {
    rust_fn_call: Option<TokenStream>,
    pre_fn_call: Option<TokenStream>,
    params: Vec<TokenStream>,
    param_lifts: Vec<TokenStream>,
    arg_metadata_calls: Vec<TokenStream>,
}

Fields
rust_fn_call: Option<TokenStream>

Tokenstream that represents the function to call

For functions, this is simple the function ident. For methods, this will be self.{method_name}.

pre_fn_call: Option<TokenStream>

Tokenstream of statements that we should have before rust_fn_call

params: Vec<TokenStream>

Parameters for the scaffolding function

param_lifts: Vec<TokenStream>

Expressions to lift the arguments in order to pass them to the exported function

arg_metadata_calls: Vec<TokenStream>

MetadataBuffer calls to build up the metadata

Implementations

impl ScaffoldingBits

fn new() -> Self

fn collect_param( &mut self, param: TokenStream, param_lift: TokenStream, metadata_builder_call: TokenStream )

fn collect_param_receiver_error( &mut self, receiver: impl ToTokens, receiver_error_msg: &str )

fn collect_params<'a>( &mut self, inputs: impl IntoIterator<Item = &'a FnArg>, receiver_error_msg: &'static str )

fn set_rust_fn_call(&mut self, rust_fn_call: TokenStream)

fn set_pre_fn_call(&mut self, pre_fn_call: TokenStream)

fn add_self_param(&mut self, param: TokenStream)

fn rust_fn_call(&self) -> TokenStream

fn gen_function_meta_static_var( &self, sig: &FnSignature, mod_path: &str ) -> Result<TokenStream>

fn gen_constructor_meta_static_var( &self, self_ident: &Ident, sig: &ConstructorSignature, mod_path: &str ) -> TokenStream

fn gen_method_meta_static_var( &self, self_ident: &Ident, sig: &FnSignature, mod_path: &str, self_is_trait: bool ) -> TokenStream